LangChainGraph/requirements.txt
# LangChain & LangGraph Learning Notebooks - Dependencies
# Install: pip install -r requirements.txt

# Core LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# LangGraph
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0

# Model Providers
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0

# Vector Stores & Embeddings (for RAG in specialty notebook)
chromadb>=0.5.0
langchain-chroma>=0.1.0

# Document Loaders
beautifulsoup4>=4.12.0
pypdf>=4.0.0

# Environment & Utilities
python-dotenv>=1.0.0
jupyter>=1.0.0
ipykernel>=6.0.0

# Observability (optional - for specialty notebook)
langsmith>=0.1.0
